Inversion of logic - specify dependencies in Varnish
The backend is now responsibile for sending a 'Dependencies' header, which is a
list of key/value pairs of table name -> row ID. These are essentialy memcached
keys, but an object can have multiple keys. To ban an item, we can now match
this set of dependencies with a regular expression.